[Bugfix] use bigInt.fromArray to parse indicies.
Discovered a bug where using '-' as a valid alphabet character caused
issues when it appeared in the most significant digit.  The bigInt
constructor interpreted it as a negation and changed the meaning of the
value to something unintended.  By switching to bigInt.fromArray and
translating each character's value ourselves, we avoid this issue.